Server/flask-server/requirements.txt
click==7.1.2
eunjeon==0.4.0
Flask==1.1.4
Flask-Cors==3.0.10
Flask-MySQLdb==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.2.0
MarkupSafe==2.0.1
mecab-python3==1.0.6
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.3
pandas==2.0.1
protobuf==3.20.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tzdata==2023.3
Werkzeug==1.0.1
